Move api files, fix project list in cmp
