+++ modules/openapi/api/apis/dop/cmdb_app_publish_item_relation_get.go
@@ -12,22 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package core_services
+package dop
 
 import (
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/modules/openapi/api/apis"
 )
 
-var APP_PUBLISH_ITEM_RELATION_UPDATE = apis.ApiSpec{
-	Path:         "/api/applications/<application>/actions/update-publish-item-relations",
-	BackendPath:  "/api/applications/<application>/actions/update-publish-item-relations",
+var APP_PUBLISH_ITEM_RELATION_GET = apis.ApiSpec{
+	Path:         "/api/applications/<application>/actions/get-publish-item-relations",
+	BackendPath:  "/api/applications/<application>/actions/get-publish-item-relations",
 	Host:         "dop.marathon.l4lb.thisdcos.directory:9527",
 	Scheme:       "http",
-	Method:       "POST",
+	Method:       "GET",
 	CheckLogin:   true,
 	CheckToken:   true,
-	RequestType:  apistructs.UpdateAppPublishItemRelationRequest{},
-	ResponseType: apistructs.UpdateAppPublishItemRelationResponse{},
-	Doc:          "summary: 更新应用发布关联",
+	ResponseType: apistructs.QueryAppPublishItemRelationResponse{},
+	Doc:          "summary: 查询应用发布关联",
 }
